pyproject.toml
[tool.poetry]
name = "orderdata-pipeline"
version = "0.1.0"
description = ""
authors = ["JuliManhupli <juliamangup@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.13"
apache-airflow = "^2.10.5"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
